+++ src/subnet/cli.py
@@ -12,7 +12,7 @@ from substrateinterface import Keypair
 
 from src.subnet.validator.challenges.generator_thread import ChallengeGeneratorThread
 from src.subnet.validator.database.models.challenge_balance_tracking import ChallengeBalanceTrackingManager
-from src.subnet.validator.database.models.challenge_funds_flow import ChallengeFundsFlowManager
+from src.subnet.validator.database.models.challenge_money_flow import ChallengeMoneyFlowManager
 from src.subnet.validator.database.models.miner_discovery import MinerDiscoveryManager
 from src.subnet.validator.database.models.miner_receipt import MinerReceiptManager
 from src.subnet.validator.database.session_manager import DatabaseSessionManager, run_migrations
@@ -79,7 +79,7 @@ if __name__ == "__main__":
 
     miner_discovery_manager = MinerDiscoveryManager(session_manager)
     miner_receipt_manager = MinerReceiptManager(session_manager)
-    challenge_funds_flow_manager = ChallengeFundsFlowManager(session_manager)
+    challenge_money_flow_manager = ChallengeMoneyFlowManager(session_manager)
     challenge_balance_tracking_manager = ChallengeBalanceTrackingManager(session_manager)
 
     receipt_sync_worker = ReceiptSyncWorker(keypair, settings.NET_UID, c_client, miner_receipt_manager)
@@ -91,7 +91,7 @@ if __name__ == "__main__":
         c_client,
         weights_storage,
         miner_discovery_manager,
-        challenge_funds_flow_manager,
+        challenge_money_flow_manager,
         challenge_balance_tracking_manager,
         miner_receipt_manager,
         redis_client=redis_client,