+++ src/subnet/gateway/routes/v1/money_flow.py
@@ -4,11 +4,11 @@ from pydantic import BaseModel
 from src.subnet.protocol import NETWORK_BITCOIN, NETWORK_COMMUNE
 from src.subnet.validator.validator import Validator
 from src.subnet.gateway import get_validator, api_key_auth
-from src.subnet.gateway.services.bitcoin_funds_flow_query_api import BitcoinFundsFlowQueryApi
-from src.subnet.gateway.services.commune_funds_flow_query_api import CommuneFundsFlowQueryApi
+from src.subnet.gateway.services.bitcoin_money_flow_query_api import BitcoinMoneyFlowQueryApi
+from src.subnet.gateway.services.commune_money_flow_query_api import CommuneMoneyFlowQueryApi
 from src.subnet.gateway.helpers.reponse_formatter import format_response, ResponseType
 
-funds_flow_router = APIRouter(prefix="/v1/funds-flow", tags=["funds-flow"])
+money_flow_router = APIRouter(prefix="/v1/money-flow", tags=["money-flow"])
 
 
 class MinerMetadataRequest(BaseModel):
@@ -18,13 +18,13 @@ class MinerMetadataRequest(BaseModel):
 def select_query_api(network: str, validator: Validator):
     """Helper function to select the appropriate query API."""
     if network == NETWORK_BITCOIN:
-        return BitcoinFundsFlowQueryApi(validator)
+        return BitcoinMoneyFlowQueryApi(validator)
     elif network == NETWORK_COMMUNE:
-        return CommuneFundsFlowQueryApi(validator)
+        return CommuneMoneyFlowQueryApi(validator)
     raise HTTPException(status_code=400, detail="Invalid network.")
 
 
-@funds_flow_router.get("/{network}/get_block")
+@money_flow_router.get("/{network}/get_block")
 async def get_blocks(
     network: str,
     block_height: int = Query(..., description="Block height"),
@@ -44,7 +44,7 @@ async def get_blocks(
     return format_response(data, response_type)
 
 
-@funds_flow_router.get("/{network}/get_transaction_by_tx_id")
+@money_flow_router.get("/{network}/get_transaction_by_tx_id")
 async def get_transaction_by_tx_id(
     network: str,
     tx_id: str,
@@ -64,7 +64,7 @@ async def get_transaction_by_tx_id(
     return format_response(data, response_type)
 
 
-@funds_flow_router.get("/{network}/get_address_transactions")
+@money_flow_router.get("/{network}/get_address_transactions")
 async def get_address_transactions(
     network: str,
     address: str = Query(...),