+++ src/subnet/protocol/__init__.py
@@ -16,16 +16,16 @@ class Discovery(BaseModel):
 
 
 # Model types
-MODEL_KIND_FUNDS_FLOW = "funds_flow"
+MODEL_KIND_MONEY_FLOW = "money_flow"
 MODEL_KIND_BALANCE_TRACKING = "balance_tracking"
 
 
 def get_model_kinds():
-    return [MODEL_KIND_FUNDS_FLOW, MODEL_KIND_BALANCE_TRACKING]
+    return [MODEL_KIND_MONEY_FLOW, MODEL_KIND_BALANCE_TRACKING]
 
 
 class Challenge(BaseModel):
-    model_kind: str = Field(default=MODEL_KIND_FUNDS_FLOW)
+    model_kind: str = Field(default=MODEL_KIND_MONEY_FLOW)
     in_total_amount: Optional[int] = None
     out_total_amount: Optional[int] = None
     tx_id_last_6_chars: Optional[str] = None
@@ -35,9 +35,9 @@ class Challenge(BaseModel):
 
 
 class ChallengesResponse(BaseModel):
-    funds_flow_challenge_expected: str
+    money_flow_challenge_expected: str
     balance_tracking_challenge_expected: int
-    funds_flow_challenge_actual: Optional[str]
+    money_flow_challenge_actual: Optional[str]
     balance_tracking_challenge_actual: Optional[int]
 
 
@@ -46,17 +46,17 @@ class ChallengeMinerResponse(BaseModel):
     version: float
     graph_db: str
 
-    funds_flow_challenge_expected: str
+    money_flow_challenge_expected: str
     balance_tracking_challenge_expected: int
-    funds_flow_challenge_actual: Optional[str]
+    money_flow_challenge_actual: Optional[str]
     balance_tracking_challenge_actual: Optional[int]
 
     def get_failed_challenges(self):
-        funds_flow_challenge_passed = self.funds_flow_challenge_expected == self.funds_flow_challenge_actual
+        money_flow_challenge_passed = self.money_flow_challenge_expected == self.money_flow_challenge_actual
         balance_tracking_challenge_passed = self.balance_tracking_challenge_expected == self.balance_tracking_challenge_actual
 
         failed_challenges = 0
-        if funds_flow_challenge_passed is False:
+        if money_flow_challenge_passed is False:
             failed_challenges += 1
 
         if balance_tracking_challenge_passed is False: